+++ src/main/java/com/example/bcsd/controller/dto/resopnse/JsonResponse.java
@@ -1,4 +1,4 @@
-package com.example.bcsd.controller.dto;
+package com.example.bcsd.controller.dto.resopnse;
 
 public class JsonResponse {
     private int age;
