+++ src/main/java/com/example/bcsd/service/ArticleService.java
@@ -1,8 +1,8 @@
 package com.example.bcsd.service;
 
-import com.example.bcsd.controller.dto.ArticleCreateRequest;
-import com.example.bcsd.controller.dto.ArticleResponse;
-import com.example.bcsd.controller.dto.ArticleUpdateRequest;
+import com.example.bcsd.controller.dto.request.ArticleCreateRequest;
+import com.example.bcsd.controller.dto.resopnse.ArticleResponse;
+import com.example.bcsd.controller.dto.request.ArticleUpdateRequest;
 import com.example.bcsd.model.Article;
 import com.example.bcsd.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
